[REFACTOR] 유저 구분 로직 변경
